Homework2

+++ Posts/DAL/Repositories/PostRepo.cs
@@ -1,9 +1,8 @@
 ﻿using AutoMapper;
 using AutoMapper.QueryableExtensions;
-using BLL.Models;
+using BLL.Models.Post;
 using DAL.Interfaces;
 using Domain.Entity;
-using Domain.Entity.Attach;
 using Microsoft.EntityFrameworkCore;
 
 namespace DAL.Repositories
@@ -20,26 +19,11 @@ namespace DAL.Repositories
             _context = context;
         }
 
-        public async Task<Guid> InsertAsync(CreatePostModel entity, Dictionary<string, MetaDataModel> files)
+        public async Task<Guid> InsertAsync(CreatePostModel entity)
         {
             var dbPost = _mapper.Map<Post>(entity);
+
             await _context.Posts.AddAsync(dbPost);
-
-            foreach (var meta in files)
-            {
-                var attach = new Attach()
-                {
-                    Id = Guid.NewGuid(),
-                    Name = meta.Value.Name,
-                    MimeType = meta.Value.MimeType,
-                    FilePath = meta.Key,
-                    Size = meta.Value.Size,
-                    Post = dbPost,
-                };
-
-                await _context.Attaches.AddAsync(attach);
-            }
-
             await _context.SaveChangesAsync();
 
             return dbPost.Id;
@@ -50,20 +34,15 @@ namespace DAL.Repositories
             return await _context.Posts.AsNoTracking().ProjectTo<GetPostModel>(_mapper.ConfigurationProvider).ToListAsync();
         }
 
-        public async Task<GetPostModel> GetPost(Guid id)
+        public async Task<bool> DeleteAsync(Guid postId, Guid authorId)
         {
-            var comment = await _context.Posts.AsNoTracking().Where(x => x.Id == id)
-                .ProjectTo<GetPostModel>(_mapper.ConfigurationProvider).FirstOrDefaultAsync();
+            if (await _context.Posts.AsNoTracking().AnyAsync(x => x.Id == postId && x.AuthorId != authorId))
+                throw new Exception("Not enough rights");
 
-            return comment ?? new GetPostModel();
-        }
-
-        public async Task<bool> DeleteAsync(Guid id)
-        {
-            var post = await _context.Posts.AsNoTracking().FirstOrDefaultAsync(x => x.Id == id);
+            var post = await _context.Posts.AsNoTracking().FirstOrDefaultAsync(x => x.Id == postId);
 
             if (post == null)
-                return false;
+                throw new Exception("Post not found");
 
             _context.Posts.Remove(post);
